@@ -2,8 +2,8 @@
   "name": "loirephemerefront",
   "version": "0.0.0",
   "engines": {
-    "node": "16.2.0",
-    "npm": "7.6.0"
+    "node": "^16.10.0",
+    "npm": "^8.6.0"
   },
   "scripts": {
     "ng": "ng",
@@ -16,48 +16,54 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~12.2.16",
-    "@angular/common": "~12.2.16",
-    "@angular/compiler": "~12.2.16",
-    "@angular/core": "~12.2.16",
-    "@angular/forms": "~12.2.16",
-    "@angular/platform-browser": "~12.2.16",
-    "@angular/platform-browser-dynamic": "~12.2.16",
-    "@angular/router": "~12.2.16",
-    "@angular/service-worker": "^12.2.16",
+    "@angular/animations": "^14.2.9",
+    "@angular/common": "^14.2.9",
+    "@angular/compiler": "^14.2.9",
+    "@angular/core": "^14.2.9",
+    "@angular/forms": "^14.2.9",
+    "@angular/platform-browser": "^14.2.9",
+    "@angular/platform-browser-dynamic": "^14.2.9",
+    "@angular/router": "^14.2.9",
+    "@angular/service-worker": "^14.2.9",
+    "@ant-design/colors": "^6.0.0",
     "@auth0/angular-jwt": "^5.0.2",
-    "@ngrx/effects": "^12.0.0",
-    "@ngrx/store": "^12.0.0",
-    "@ngrx/store-devtools": "^12.0.0",
+    "@ngrx/effects": "^14.3.2",
+    "@ngrx/store": "^14.3.2",
+    "@ngrx/store-devtools": "^14.3.2",
     "@types/chart.js": "^2.9.34",
     "@types/resize-observer-browser": "^0.1.5",
     "chart.js": "^3.8.2",
     "chartjs-plugin-datalabels": "^2.0.0",
     "moment": "^2.29.1",
-    "ng-zorro-antd": "^12.0.0",
-    "ng2-charts": "^3.0.0",
-    "rxjs": "~6.6.0",
+    "ng-zorro-antd": "^14.1.1",
+    "ng2-charts": "^4.0.1",
+    "rxjs": "^7.5.7",
     "tslib": "^2.0.0",
-    "zone.js": "~0.11.4"
+    "zone.js": "^0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.2.16",
-    "@angular/cli": "~12.0.3",
-    "@angular/compiler-cli": "~12.2.16",
-    "@schematics/angular": "^11.2.11",
-    "@types/jasmine": "~3.6.0",
-    "@types/node": "^12.11.1",
-    "codelyzer": "^6.0.0",
-    "jasmine-core": ">=3.8.0",
-    "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~6.3.2",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage": "~2.0.3",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
-    "protractor": "~7.0.0",
-    "ts-node": "~8.3.0",
-    "tslint": "~6.1.0",
-    "typescript": "~4.2.4"
+    "@angular-devkit/build-angular": "^14.2.8",
+    "@angular-eslint/builder": "14.1.2",
+    "@angular-eslint/eslint-plugin": "14.1.2",
+    "@angular-eslint/eslint-plugin-template": "14.1.2",
+    "@angular-eslint/schematics": "14.1.2",
+    "@angular-eslint/template-parser": "14.1.2",
+    "@angular/cli": "^14.2.8",
+    "@angular/compiler-cli": "^14.2.9",
+    "@schematics/angular": "^14.2.8",
+    "@types/jasmine": "^4.3.0",
+    "@types/node": "^16.x",
+    "@typescript-eslint/eslint-plugin": "5.42.0",
+    "@typescript-eslint/parser": "5.42.0",
+    "eslint": "^8.27.0",
+    "jasmine-core": "^4.0.0",
+    "jasmine-spec-reporter": "^7.0.0",
+    "karma": "^6.4.1",
+    "karma-chrome-launcher": "^3.1.0",
+    "karma-coverage": "^2.2.0",
+    "karma-jasmine": "^5.1.0",
+    "karma-jasmine-html-reporter": "^2.0.0",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.8.4"
   }
 }
