@@ -1,32 +1,50 @@
-export const pastriesMock = [
+import { Pastry } from "src/app/interfaces/pastry.interface";
+import { restaurantsMock } from "./restaurants.mock";
+
+export const pastriesMock: Pastry[] = [
   {
     _id: '2345YT',
     name: '2345YT',
     price: 24,
     description: '2345YT',
-    imageUrl: '2345YT',
+    imageUrl: null,
     type: 'pastry',
     stock: 3,
+    displaySequence: 3,
+    hidden: false,
     ingredients: [''],
+    restaurant: restaurantsMock[0],
+    createdAt: '2021-07-22T09:49:07.324+00:00',
+    updatedAt: '2021-07-22T09:49:07.324+00:00',
   },
   {
     _id: '2349YT',
     name: '2345YT',
     price: 24,
     description: '2345YT',
-    imageUrl: '2345YT',
+    imageUrl: null,
     type: 'pastry',
     stock: 3,
+    displaySequence: 4,
+    hidden: false,
     ingredients: [''],
+    restaurant: restaurantsMock[0],
+    createdAt: '2021-07-22T09:49:07.324+00:00',
+    updatedAt: '2021-07-22T09:49:07.324+00:00',
   },
   {
     _id: '2348YT',
     name: '2345YT',
     price: 24,
     description: '2345YT',
-    imageUrl: '2345YT',
+    imageUrl: null,
     type: 'pastry',
     stock: 3,
+    displaySequence: 5,
+    hidden: false,
     ingredients: [''],
+    restaurant: restaurantsMock[0],
+    createdAt: '2021-07-22T09:49:07.324+00:00',
+    updatedAt: '2021-07-22T09:49:07.324+00:00',
   },
 ];
