@@ -1,25 +1,52 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { SharedModule } from '../../shared/shared.module';
-import { AdminComponent } from './components/admin/admin.component';
-import { CardComponent } from './components/card/card.component';
+import { AdminComponent } from './components/admin.component';
+import { CommandCardComponent } from './components/command-card/command-card.component';
 import { EffectsModule } from '@ngrx/effects';
 import { StoreModule } from '@ngrx/store';
 import { AdminEffects } from './store/admin.effects';
 import { adminFeatureKey, reducer } from './store/admin.reducer';
 import { StatsComponent } from './components/stats/stats.component';
 import { NgChartsModule } from 'ng2-charts';
-import { RouterModule } from '@angular/router';
-import { FormsModule } from '@angular/forms';
+import { AdminRoutingModule } from './admin-routing.module';
+import { CommandsComponent } from './components/commands/commands.component';
+import { MenuComponent } from './components/menu/menu.component';
+import { NewPastryModalComponent } from 'src/app/modules/admin/components/new-pastry-modal/new-pastry-modal.component';
+import { EditPastryModalComponent } from 'src/app/modules/admin/components/edit-pastry-modal/edit-pastry-modal.component';
+import { PastryFormComponent } from 'src/app/modules/admin/components/pastry-form/pastry-form.component';
+import { DragDropModule } from '@angular/cdk/drag-drop';
+import { UsersComponent } from 'src/app/modules/admin/components/users/users.component';
+import { NewUserModalComponent } from 'src/app/modules/admin/components/new-user-modal/new-user-modal.component';
+import { BarChartComponent } from './components/charts/bar-chart/bar-chart.component';
+import { PieChartComponent } from './components/charts/pie-chart/pie-chart.component';
+import { AssociationComponent } from './components/menu/association/association.component';
+import { SequenceComponent } from './components/menu/sequence/sequence.component';
+
 
 @NgModule({
-  declarations: [AdminComponent, StatsComponent, CardComponent],
+  declarations: [
+    AdminComponent,
+    CommandsComponent,
+    StatsComponent,
+    CommandCardComponent,
+    MenuComponent,
+    NewPastryModalComponent,
+    EditPastryModalComponent,
+    PastryFormComponent,
+    UsersComponent,
+    NewUserModalComponent,
+    BarChartComponent,
+    PieChartComponent,
+    AssociationComponent,
+    SequenceComponent,
+  ],
   imports: [
     CommonModule,
     SharedModule,
     NgChartsModule,
-    FormsModule,
-    RouterModule,
+    AdminRoutingModule,
+    DragDropModule,
     EffectsModule.forFeature([AdminEffects]),
     StoreModule.forFeature(adminFeatureKey, reducer),
   ],
