@@ -1,7 +1,6 @@
 import { createSelector } from '@ngrx/store';
 import { AppState } from 'src/app/store/app.state';
 import { AdminState } from './admin.reducer';
-import { commandsMock } from 'src/app/mocks/commands.mock';
 
 export const selectFeature = (state: AppState) => state.admin;
 
@@ -53,4 +52,64 @@ export const selectTotalPayedCommands = createSelector(
 export const selectIsLoading = createSelector(
   selectFeature,
   (state: AdminState) => state.loading
+);
+
+export const selectIsStatsLoading = createSelector(
+  selectFeature,
+  (state: AdminState) => state.statsLoading
+);
+
+export const selectIsSavingPastry = createSelector(
+  selectFeature,
+  (state: AdminState) => state.isSavingPastry
+);
+
+export const selectIsMovingPastry = createSelector(
+  selectFeature,
+  (state: AdminState) => state.isMovingPastry
+);
+
+export const selectAllPastries = createSelector(
+  selectFeature,
+  (state: AdminState) => state.allPastries
+);
+
+export const selectIsNameValidating = createSelector(
+  selectFeature,
+  (state: AdminState) => state.isNameValidating
+);
+
+export const selectPastryNameError = createSelector(
+  selectFeature,
+  (state: AdminState) => state.pastryNameError
+);
+
+export const selectMenuModalOpened = createSelector(
+  selectFeature,
+  (state: AdminState) => state.menuModalOpened
+);
+
+export const selectPastryNameDeactivated = createSelector(
+  selectFeature,
+  (state: AdminState) => state.pastryNameDeactivated
+);
+
+export const selectEditingPastry = createSelector(
+  selectFeature,
+  (state: AdminState) => state.editingPastry
+);
+
+export const selectUsers = createSelector(
+  selectFeature,
+  (state: AdminState) => state.users
+);
+
+export const selectUserEmailError = createSelector(
+  selectFeature,
+  (state: AdminState) => state.userEmailError
+);
+
+export const selectIsEmailValidating = createSelector(
+  selectFeature,
+  (state: AdminState) => state.isEmailValidating
 );
