@@ -1,15 +1,13 @@
-import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
+import { Component, EventEmitter, Input, Output } from '@angular/core';
 
 @Component({
   selector: 'app-order-success-modal',
   templateUrl: './order-success-modal.component.html',
   styleUrls: ['./order-success-modal.component.scss'],
 })
-export class OrderSuccessModalComponent implements OnInit {
+export class OrderSuccessModalComponent {
   @Input() commandReference: string = '';
-  @Output() onClickCancel = new EventEmitter<string>();
+  @Output() clickCancel = new EventEmitter<string>();
 
-  constructor() {}
-
-  ngOnInit(): void {}
+  constructor() { }
 }
