@@ -1,14 +1,12 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { HomeComponent } from './components/home/home.component';
+import { HomeComponent } from './components/home.component';
 import { SharedModule } from '../../shared/shared.module';
 import { homeFeatureKey, reducer } from './store/home.reducer';
 import { StoreModule } from '@ngrx/store';
 import { EffectsModule } from '@ngrx/effects';
 import { HomeEffects } from './store/home.effects';
-import { CardComponent } from './components/card/card.component';
 import { OrderFooterComponent } from './components/order-footer/order-footer.component';
-import { FormsModule } from '@angular/forms';
 import { OrderModalComponent } from './components/order-modal/order-modal.component';
 import { OrderNameModalComponent } from './components/order-name-modal/order-name-modal.component';
 import { OrderSuccessModalComponent } from './components/order-success-modal/order-success-modal.component';
@@ -18,7 +16,6 @@ import { OrderErrorModalComponent } from './components/order-error-modal/order-e
   declarations: [
     HomeComponent,
     OrderFooterComponent,
-    CardComponent,
     OrderNameModalComponent,
     OrderSuccessModalComponent,
     OrderErrorModalComponent,
@@ -27,7 +24,6 @@ import { OrderErrorModalComponent } from './components/order-error-modal/order-e
   imports: [
     CommonModule,
     SharedModule,
-    FormsModule,
     EffectsModule.forFeature([HomeEffects]),
     StoreModule.forFeature(homeFeatureKey, reducer),
   ],
