@@ -4,7 +4,52 @@ import { LoginState } from './login.reducer';
 
 export const selectFeature = (state: AppState) => state.login;
 
-export const selectToken = createSelector(
+export const selectUserEmailError = createSelector(
   selectFeature,
-  (state: LoginState) => state.token
+  (state: LoginState) => state.userEmailError
+);
+
+export const selectIsEmailValidating = createSelector(
+  selectFeature,
+  (state: LoginState) => state.isEmailValidating
+);
+
+export const selectUserAuthError = createSelector(
+  selectFeature,
+  (state: LoginState) => state.userAuthError
+);
+
+export const selectUser = createSelector(
+  selectFeature,
+  (state: LoginState) => state.user
+);
+
+export const selectUserRestaurants = createSelector(
+  selectFeature,
+  (state: LoginState) => state.userRestaurants
+);
+
+export const selectLoading = createSelector(
+  selectFeature,
+  (state: LoginState) => state.loading
+);
+
+export const selectCode2 = createSelector(
+  selectFeature,
+  (state: LoginState) => state.code2
+);
+
+export const selectConfirmationModalOpened = createSelector(
+  selectFeature,
+  (state: LoginState) => state.confirmationModalOpened
+);
+
+export const selectRecoverModalOpened = createSelector(
+  selectFeature,
+  (state: LoginState) => state.recoverModalOpened
+);
+
+export const selectPasswordChanged = createSelector(
+  selectFeature,
+  (state: LoginState) => state.passwordChanged
 );
